+++ src/sparseml/modifiers/quantization_legacy/utils/quantization_scheme.py
@@ -30,7 +30,7 @@ try:
 except Exception:
     torch_quantization = None
 
-from sparseml.modifiers.quantization.utils.fake_quant_wrapper import FakeQuantizeWrapper
+from sparseml.modifiers.quantization_legacy.utils.fake_quant_wrapper import FakeQuantizeWrapper
 
 
 __all__ = [
